first set of ratings are added successfully

+++ db/migrate/20190719193351_create_rating.rb
@@ -1,7 +1,7 @@
 class CreateRating < ActiveRecord::Migration[5.2]
   def change
     create_table :ratings do |t|
-      t.integer :rating, null: false
+      t.float :rating, null: false
       
       t.belongs_to :card
       t.belongs_to :rating_system
